+++ src/legacy/core_plugins/console/public/np_ready/application/models/sense_editor/sense_editor.ts
@@ -78,7 +78,7 @@ export class SenseEditor {
     } else {
       curRow = rowOrPos as number;
     }
-    const maxLines = this.coreEditor.getValue().split('\n').length;
+    const maxLines = this.coreEditor.getLineCount();
     for (; curRow < maxLines - 1; curRow++) {
       if (this.parser.isStartRequestRow(curRow, this.coreEditor)) {
         break;
